scripts/sick_mine_llm/requirements.txt
langchain>=0.0.200
langchain-community>=0.1.0
streamlit>=1.24.0
faiss-cpu>=1.7.4
openai>=0.28.0
python-dotenv>=1.0.0
langchain-nvidia-ai-endpoints
PyPDF2>=3.0.0
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.2
